refactor(utils/xyz_dododo): 创建文件夹 `src/utils` 并将 部分代码 移入其中 & 更新导入
`utils/` 统一管理工具函数
