+++ src/plugins/chatSpawn.ts
@@ -9,7 +9,7 @@ import {
 } from './main'
 import { type CommandInfo, commandManager, Command } from '../command'
 import { Dimension, Vector3, world, type Player } from '@minecraft/server'
-import {xyz_dododo} from "../lib/xboyPackage/xyz_dododo";
+import {xyz_dododo} from "../utils/xyz_dododo";
 
 const overworld = world.getDimension("overworld");
