+++ src/plugins/task.ts
@@ -4,7 +4,7 @@ import {simulatedPlayers, testWorldLocation} from './main'
 import SIGN from '../lib/xboyPackage/YumeSignEnum'
 import type { EntityHealthComponent, Vector3 } from '@minecraft/server'
 import { system, world } from '@minecraft/server'
-import { getEntitiesNear, getPlayerNear } from '../lib/xboyPackage/Util'
+import { getEntitiesNear, getPlayerNear } from '../utils/Util'
 
 // @ts-ignore
 const simulatedPlayerStates : ({ "str-SimPlayer.id": { o: Vector3 }}) = {}
